Gets rid of the awful classes
