+++ plugins/catalog-graphql/src/graphql-tools/index.ts
@@ -13,4 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './field-parser';
+export * from './field-builder';
+export * from './filters';
+export * from './object-builder';
